Move the FXL inline scripts to the bundled scripts
