+++ Sources/Navigator/EPUB/Assets/Static/scripts/readium-reflowable.js
@@ -1709,7 +1709,7 @@ __webpack_require__.r(__webpack_exports__);
 
  // Public API used by the navigator.
 
-window.readium = {
+__webpack_require__.g.readium = {
   // utils
   scrollToId: _utils__WEBPACK_IMPORTED_MODULE_1__.scrollToId,
   scrollToPosition: _utils__WEBPACK_IMPORTED_MODULE_1__.scrollToPosition,
@@ -8057,6 +8057,18 @@ module.exports = function sign(number) {
 /******/ 				}
 /******/ 			}
 /******/ 		};
+/******/ 	})();
+/******/ 	
+/******/ 	/* webpack/runtime/global */
+/******/ 	(() => {
+/******/ 		__webpack_require__.g = (function() {
+/******/ 			if (typeof globalThis === 'object') return globalThis;
+/******/ 			try {
+/******/ 				return this || new Function('return this')();
+/******/ 			} catch (e) {
+/******/ 				if (typeof window === 'object') return window;
+/******/ 			}
+/******/ 		})();
 /******/ 	})();
 /******/ 	
 /******/ 	/* webpack/runtime/hasOwnProperty shorthand */
