+++ Sources/Navigator/EPUB/Assets/fxl-spread-two.html
@@ -1,8 +1,6 @@
 <!DOCTYPE html>
 <html>
 <head>
-  <script type="text/javascript" src="/r2-navigator/epub/scripts/readium-fixed.js"></script>
-  <script type="text/javascript" src="/r2-navigator/epub/scripts/readium-fixed-wrapper.js"></script>
   <style>
   body, html {
     height: 100%;
@@ -72,103 +70,7 @@
   <div id="viewport-center" class="viewport">
     <iframe id="page-center" class="page"></iframe>
   </div>
-
-  <script type="text/javascript">
-    var spread = (function() {
-      var _viewportLeft = document.getElementById('viewport-left');
-      var _viewportRight = document.getElementById('viewport-right');
-      var _viewportCenter = document.getElementById('viewport-center');
-
-      var _pages = {
-        'left': FixedPage('page-left'),
-        'right': FixedPage('page-right'),
-        'center': FixedPage('page-center')
-      };
-
-      function forEachPage(callback) {
-        for (position in _pages) {
-          callback(_pages[position]);
-        }
-      }
-
-      function getPageWithHref(href) {
-        for (position in _pages) {
-          var page = _pages[position];
-          if (page.link?.href === href) {
-            return page;
-          }
-        }
-        return null;
-      }
-
-      // Public API called from Swift.
-      return {
-        // Loads resources in the spread.
-        'load': function(resources) {
-          forEachPage(function(page) {
-            page.reset();
-            page.hide();
-          });
-
-          function loaded() {
-            if (_pages.left.isLoading || _pages.right.isLoading || _pages.center.isLoading) {
-              return;
-            }
-            webkit.messageHandlers.spreadLoaded.postMessage({});
-          }
-
-          for (const i in resources) {
-            const resource = resources[i];
-            const page = _pages[resource.page];
-            if (page) {
-              page.show();
-              page.load(resource, loaded);
-            }
-          }
-        },
-
-        // Evaluates a JavaScript in the context of a resource.
-        // If the href is '#' or empty, then the script is executed on all the pages.
-        'eval': function(href, script) {
-          if (href === '#' || href === '') {
-            forEachPage(function(page) {
-              page.eval(script);
-            });
-          } else {
-            var page = getPageWithHref(href);
-            if (page) {
-              return page.eval(script);
-            }
-          }
-        },
-
-        // Updates the available viewport to display the resources.
-        'setViewport': function(viewportSize, safeAreaInsets) {
-          viewportSize.width /= 2;
-
-          _pages.left.setViewport(viewportSize, {
-            top: safeAreaInsets.top,
-            right: 0,
-            bottom: safeAreaInsets.bottom,
-            left: safeAreaInsets.left
-          });
-
-          _pages.right.setViewport(viewportSize, {
-            top: safeAreaInsets.top,
-            right: safeAreaInsets.right,
-            bottom: safeAreaInsets.bottom,
-            left: 0
-          });
-
-          _pages.center.setViewport(viewportSize, {
-            top: safeAreaInsets.top,
-            right: 0,
-            bottom: safeAreaInsets.bottom,
-            left: 0
-          });
-        }
-      }
-    })();
-  </script>
+  
+  <script type="text/javascript" src="/r2-navigator/epub/scripts/readium-fixed-wrapper-two.js"></script>
 </body>
 </html>
